change mod name, re-init go.mod
