+++ .git-ignore
@@ -1,0 +1,2 @@
+# executable file
+htsget-refserver
