+++ go.mod
@@ -1,21 +1,5 @@
-module github.com/david-xliu/htsget_refserver
+module github.com/david-xliu/htsget-refserver
 
 go 1.13
 
-require (
-	github.com/cosiner/argv v0.0.1 // indirect
-	github.com/go-chi/chi v4.0.2+incompatible
-	github.com/go-delve/delve v1.3.1 // indirect
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/mattn/go-colorable v0.1.2 // indirect
-	github.com/mattn/go-isatty v0.0.9 // indirect
-	github.com/mattn/go-runewidth v0.0.4 // indirect
-	github.com/peterh/liner v1.1.0 // indirect
-	github.com/sirupsen/logrus v1.4.2 // indirect
-	github.com/spf13/cobra v0.0.5 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-	go.starlark.net v0.0.0-20190919145610-979af19b165c // indirect
-	golang.org/x/arch v0.0.0-20190909030613-46d78d1859ac // indirect
-	golang.org/x/crypto v0.0.0-20190911031432-227b76d455e7 // indirect
-	golang.org/x/sys v0.0.0-20190919044723-0c1ff786ef13 // indirect
-)
+require github.com/go-chi/chi v4.0.2+incompatible
