+++ main.go
@@ -1,11 +1,7 @@
 package main
 
 import (
-	"encoding/json"
 	"net/http"
-	"path/filepath"
-	"strconv"
-	"strings"
 
 	"github.com/david-xliu/htsget-refserver/handler"
 )
